requirements.txt
numpy
opencv-python
torch
Pillow
trimesh
rembg
transformers
onnxruntime
